+++ packages/xchain-bitcoincash/__mocks__/api.ts
@@ -1,0 +1,8 @@
+import nock from 'nock'
+import { AddressDetails } from '../src/types'
+
+export const mock_getBalance = (url: string, address: string, result: AddressDetails) => {
+  nock(url)
+    .get(`/address/details/${address}`)
+    .reply(200, result)
+}